Implement ResponseEntity for proper REST API responses

+++ src/main/java/com/example/practice/controller/StudentController.java
@@ -3,6 +3,8 @@ package com.example.practice.controller;
 import com.example.practice.entity.Student;
 import com.example.practice.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -17,31 +19,51 @@ public class StudentController {
     private StudentService studentService;
 
     @GetMapping
-    public List<Student> getAllStudent()
+    public ResponseEntity<List<Student>> getAllStudent()
     {
-        return studentService.getAllStudent();
+        List <Student> students=studentService.getAllStudent();
+        return ResponseEntity.ok(students);
     }
 
     @PostMapping
-    public Student createStudent(@RequestBody Student student)
+    public ResponseEntity<Student> createStudent(@RequestBody Student student)
     {
-        return studentService.createStudent(student);
+        Student savedStudent= studentService.createStudent(student);
+        return new ResponseEntity<>(savedStudent, HttpStatus.CREATED);
+        //Alternate
+        //return ResponseEntity.status(201).body(savedStudent);
     }
     @GetMapping("/{id}")
-    public Student getStudentById(@PathVariable int id)
+    public ResponseEntity<Student> getStudentById(@PathVariable int id)
     {
-       return studentService.getStudentById(id);
+        Student student=studentService.getStudentById(id);
+        if(student==null)
+        {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+       return new ResponseEntity<>(student,HttpStatus.FOUND);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable int id)
+    public ResponseEntity<String> deleteById(@PathVariable int id)
     {
+        Student student=studentService.getStudentById(id);
+        if(student==null)
+        {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
        studentService.deleteStudent(id);
+        return new ResponseEntity<>("Student deleted successfully",HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public Student updateStudent(@PathVariable int id,@RequestBody Student updatedStudent)
+    public ResponseEntity<Student> updateStudent(@PathVariable int id,@RequestBody Student student)
     {
-        return studentService.updateStudent(id,updatedStudent);
+        Student updatedStudent=studentService.updateStudent(id,student);
+        if(updatedStudent==null)
+        {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(updatedStudent,HttpStatus.OK);
     }
 }
